refactor: code review improvements
